Change int to either NSUInteger (for indices) or NSInteger Using integerValue instead of intValue Update specs to follow.

+++ Classes/NSArray+ObjectiveSugar.m
@@ -28,7 +28,7 @@
     }];
 }
 
-- (void)eachWithIndex:(void (^)(id object, int index))block {
+- (void)eachWithIndex:(void (^)(id object, NSUInteger  index))block {
     [self enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
         block(obj, idx); 
     }];
